Add files via upload
Pd patch for the Nebulae that provides a model for the ondes Martenot's Palme diffuseur. It has a dynamic number of sympathetic strings based on assignable equal tempered tuning. The user can choose feedback amount (overlap), resonance (window), a reduction in the number of strings (blend), the interval--octave, tritave, quartive, quintive, etc.--to be divided (size), the number of equal steps (division), a number of steps in the scale to be selected from the equal divisions (density), root transposition (pitch), and overdrive amplification (speed).

This is a working version of in progress software. Updates will provide better ranges and functions, possibly changing some of the input parameters.

The second audio output provides audio clues as to where the values are set and will beep when certain knobs are tweaked.